refactor(cli): move format_bytes to output.rs, add format_duration
Centralizes formatting helpers in the output module so both
synchronize.rs and the upcoming TextProgress can use them.
